+++ continuous/KraskovStoegbauerGrassberger_test.go
@@ -19,9 +19,9 @@ func TestKraskovStoegbauerGrassberger1Independent(t *testing.T) {
 
 	var xy [][]float64
 
-	xIndex := make([]int, 1, 1)
+	xIndex := make([]int64, 1, 1)
 	xIndex[0] = 0
-	yIndex := make([]int, 1, 1)
+	yIndex := make([]int64, 1, 1)
 	yIndex[0] = 1
 
 	for i := 0; i < N; i++ {
@@ -41,7 +41,7 @@ func TestKraskovStoegbauerGrassberger1Gaussian(t *testing.T) {
 	rand.Seed(1)
 
 	N := 1000
-	k := 20
+	k := int64(20)
 
 	r := 0.9 // co-variance
 	mi := 0.0
@@ -55,9 +55,9 @@ func TestKraskovStoegbauerGrassberger1Gaussian(t *testing.T) {
 
 		var xy [][]float64
 
-		xIndex := make([]int, 1, 1)
+		xIndex := make([]int64, 1, 1)
 		xIndex[0] = 0
-		yIndex := make([]int, 1, 1)
+		yIndex := make([]int64, 1, 1)
 		yIndex[0] = 1
 
 		for i := 0; i < N; i++ {
@@ -83,9 +83,9 @@ func TestKraskovStoegbauerGrassberger2Independent(t *testing.T) {
 
 	var xy [][]float64
 
-	xIndex := make([]int, 1, 1)
+	xIndex := make([]int64, 1, 1)
 	xIndex[0] = 0
-	yIndex := make([]int, 1, 1)
+	yIndex := make([]int64, 1, 1)
 	yIndex[0] = 1
 
 	for i := 0; i < N; i++ {
@@ -105,7 +105,7 @@ func TestKraskovStoegbauerGrassberger2Gaussian(t *testing.T) {
 	rand.Seed(2)
 
 	N := 1000
-	k := 30
+	k := int64(30)
 
 	r := 0.9 // co-variance
 
@@ -120,9 +120,9 @@ func TestKraskovStoegbauerGrassberger2Gaussian(t *testing.T) {
 
 		var xy [][]float64
 
-		xIndex := make([]int, 1, 1)
+		xIndex := make([]int64, 1, 1)
 		xIndex[0] = 0
-		yIndex := make([]int, 1, 1)
+		yIndex := make([]int64, 1, 1)
 		yIndex[0] = 1
 
 		for i := 0; i < N; i++ {
